Update to rustc 1.0.0-nightly (44a287e6e 2015-01-08 17:03:40 -0800)

+++ src/enums.rs
@@ -11,8 +11,8 @@ pub use self::Vg_TCheckClientRequest::*;
 pub use self::Vg_DRDClientRequest::*;
 
 macro_rules! VG_USERREQ_TOOL_BASE ( ($a:expr, $b:expr) => (
-    ((($a as int) & 0xff) << 24)
-  | ((($b as int) & 0xff) << 16)));
+    ((($a as isize) & 0xff) << 24)
+  | ((($b as isize) & 0xff) << 16)));
 
 #[repr(C)]
 pub enum Vg_ClientRequest {
